Exposed public api for creating mock settings
Framework integrators need to be able to create instances of MockCreationSettings objects without providing entire custom implementation of this interface. The latter approach is not viable because MockCreationSettings can grow with new methods when we add more configurability to mocks. When that happens, all clients who have custom implementation of MockCreationSettings will break. To avoid this risk, this change offers public API to create instances of MockCreationSettings.